refato code, delete bookCartComponent. use bookComponent for bookCartComponent and bookComponent
