+++ bookApp/src/app/book/book.component.html
@@ -4,5 +4,5 @@
     <h5 class="card-title">{{book.title}}</h5>
     <p>{{book.price}}€</p>
   </div>
-  <button (click)="addToCart(book)" class="btn btn-primary">add to cart</button>
+  <button (click)="eventButton.emit('complete')"  class="btn btn-primary">{{nameButton}}</button>
 </div>
