+++ bookApp/src/app/books/books.component.html
@@ -8,6 +8,6 @@
 
 <div class="row align">
   <div *ngFor="let book of books | filter : searchText" class="margin">
-    <app-book [book] = "book" ></app-book>
+    <app-book [book] = "book" (eventButton) = "addToCart(book)" nameButton = "add to cart"></app-book>
   </div>
 </div>
