+++ bookApp/src/app/books/books.component.ts
@@ -1,6 +1,7 @@
 import { Component , OnInit} from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { BooksService } from './books.service';
+import {Globals} from "../globals/globals";
 
 @Component({
   selector: 'app-books',
@@ -9,20 +10,22 @@ import { BooksService } from './books.service';
   styleUrls: ['./books.component.css']
 })
 export class BooksComponent implements OnInit{
-  constructor(private http: HttpClient , private booksService: BooksService){
+  constructor(private http: HttpClient , private booksService: BooksService, private globals : Globals){
 
   }
+
+  private books : Object;
 
   ngOnInit(){
-    this.showConfig();
-  }
-
-  books : Object;
-
-  showConfig() {
     this.booksService.getBooks()
       .subscribe((data) =>
         this.books = data
       );
   }
+
+  addToCart(book) {
+    this.globals.cart.push(book);
+  }
+
+
 }
