--- bookApp/src/app/cart/book/bookCart.component.html
@@ -1,8 +1,0 @@
-<div class="card" style="width: 18rem;">
-  <img class="card-img-top" [src]="book.cover" alt="cover of the book">
-  <div class="card-body">
-    <h5 class="card-title">{{book.title}}</h5>
-    <p>{{book.price}}€</p>
-  </div>
-  <button (click)="removeToCart(index)" class="btn btn-primary">remove from cart</button>
-</div>
