+++ bookApp/src/app/offer/offer.spec.ts
@@ -5,7 +5,7 @@ import { HttpClientTestingModule, HttpTestingController } from '@angular/common/
 //component
 import { AppComponent } from '../app.component';
 import {BooksComponent} from "../books/books.component";
-import {BookComponent} from "../books/book/book.component";
+import {BookComponent} from "../book/book.component";
 
 //service
 import {OfferService} from "./offer.service";
